move rustc_type_ir Term things to term_kind.rs
