+++ compiler/rustc_type_ir/src/lib.rs
@@ -52,6 +52,7 @@ mod pattern;
 mod predicate;
 mod predicate_kind;
 mod region_kind;
+mod term_kind;
 mod ty;
 mod ty_info;
 mod ty_kind;
@@ -81,6 +82,7 @@ pub use predicate_kind::*;
 pub use region_kind::*;
 pub use rustc_ast_ir::{FloatTy, IntTy, Movability, Mutability, Pinnedness, UintTy};
 use rustc_type_ir_macros::GenericTypeVisitable;
+pub use term_kind::*;
 pub use ty::{Alias, *};
 pub use ty_info::*;
 pub use ty_kind::*;
